# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/count/course/subscription": {
            "get": {
                "description": "Get Subscribed user count for certian course",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get User Count for certian course",
                "parameters": [
                    {
                        "description": "Get User Count For Certain Course",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.GetUserCountForCertianCourseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/subscription/edit": {
            "put": {
                "description": "Update a subscription",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a Subscription",
                "operationId": "update-subscription",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course ID",
                        "name": "course_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Subscription Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateSubscriptionsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/subscription/get": {
            "get": {
                "description": "Retrieves a subscription by user ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a subscription By User",
                "operationId": "GetSubscription",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "UserID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/subscriptions/user": {
            "get": {
                "description": "List subscriptions for a given user ID",
                "produces": [
                    "application/json"
                ],
                "summary": "List subscriptions",
                "operationId": "listSubscriptions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "course_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/tokens/renew_access": {
            "post": {
                "description": "Renew access token using refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Renew Access Token",
                "operationId": "renew-access-token",
                "parameters": [
                    {
                        "description": "Refresh Token",
                        "name": "refresh_token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/userStatus/edit": {
            "put": {
                "description": "Update the status of a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update an userStatus",
                "operationId": "update-userStatus",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Status ID",
                        "name": "status_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update User Status Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateUserStatusByAdminRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.GetUserCountForCertianCourseRequest": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "course_id": {
                    "type": "integer"
                }
            }
        },
...
        "api.UpdateSubscriptionsRequest": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "course_id": {
                    "type": "integer"
                },
                "pending": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "api.UpdateUserStatusByAdminRequest": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "pending": {
                    "type": "boolean"
                },
                "status_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}